reorganised code across modules
